Reshuffled PBookTyper.py versions
