+++ pandaclient/PBookTyper.py
@@ -1,5 +1,13 @@
 """
-pbook CLI — PanDA task bookkeeper with typer-based shell autocompletion.
+pbook CLI — PanDA task bookkeeper.
+
+Each command below is defined exactly once, as a Typer command using the
+Annotated[...] parameter style. Because the Typer/Click metadata lives in the
+annotation rather than in the default value, these functions remain ordinary
+callables with ordinary defaults - the same function is used to build the
+`pbook <command> --flag ...` CLI (with real shell completion) *and* is placed
+directly into the interactive REPL namespace (`>>> command(...)`), with no
+separate REPL-only copy of the command's logic, docstring, or option list.
 """
 
 from __future__ import annotations
@@ -14,12 +22,21 @@ import signal
 import sys
 import tempfile
 from concurrent.futures import ThreadPoolExecutor
-from inspect import signature
-from typing import Optional
+from inspect import Parameter, signature
+from typing import (
+    Annotated,
+    Literal,
+    Optional,
+    Union,
+    get_args,
+    get_origin,
+    get_type_hints,
+)
 
 import typer
 from rich import box
 from rich.console import Console
+from rich.markup import escape as _esc
 from rich.table import Table
 
 from pandaclient import Client, PandaToolsPkgInfo
@@ -31,25 +48,78 @@ _history_file: Optional[str] = None
 _fork_child_pid: Optional[int] = None
 _setup_done: bool = False
 _ctx_state: dict = {}
+_core = None
+_core_inited: bool = False
+
+help_text = """
+PanDA task bookkeeper. Run without arguments for interactive mode.
+
+$ pbook [options] # interactive mode
+$ pbook [options] command [args] [kwargs] # batch mode
+
+The same command can be executed in interactive mode:
+
+$ pbook
+>>> command(*args, **kwargs)
+
+or in batch mode:
+
+$ pbook command arg1 arg2 ... argN --kwarg1=value1 --kwarg2=value2 ... --kwargN=valueN
+$ pbook command arg1 arg2 ... argN kwarg1=value1 kwarg2=value2 ... kwargN=valueN
+Please note that the latter option is kept for backward compatibility, but we plan to drop it in the future. 
+
+E.g.
+
+$ pbook
+>>> show(123, format='long')
+
+is equivalent to
+
+$ pbook show 123 --format='long'
+$ pbook show 123 format='long'
+
+If arg or value is a list in interactive mode, it is represented as a comma-separate list in batch mode. E.g.
+to kill three tasks in interactive mode:
+
+$ pbook
+>>> kill([123, 456, 789])
+
+or in batch mode:
+
+$ pbook kill 123,456,789
+
+To see the list of commands and help of each command,
+
+$ pbook
+>>> help()
+>>> help(command_name)
+
+or
+
+$ pbook help
+$ pbook command_name --help
+"""
 
 app = typer.Typer(
     name="pbook",
-    help="PanDA task bookkeeper. Run without arguments for interactive mode.",
+    help=help_text,
     invoke_without_command=True,
     no_args_is_help=False,
+    context_settings={"help_option_names": ["-h", "--help"]},
 )
 
 # ─── Utilities ────────────────────────────────────────────────────────────────
 
 
 def _parallel(func, items):
-    # Parallel execution in a thread pool of 8 threads, for example when the user wants to act on a list of task IDs
     with ThreadPoolExecutor(8) as pool:
         return list(pool.map(func, items))
 
 
-def _parse_ids(raw: str):
-    """'all' → str, '42' → int, '1,2,3' → [int,...]."""
+def _parse_ids(raw):
+    """'all' -> 'all'; '42' -> 42; '1,2,3' -> [1,2,3]; anything not a string passes through as-is."""
+    if not isinstance(raw, str):
+        return raw
     if raw == "all":
         return "all"
     parts = raw.split(",")
@@ -119,8 +189,28 @@ def _make_core(verbose: bool = False):
 
 
 def _get_core():
+    """Return the (memoized) core for this process, uninitialized."""
+    global _core
     _setup()
-    return _make_core(_ctx_state.get("verbose", False))
+    if _core is None:
+        _core = _make_core(_ctx_state.get("verbose", False))
+    return _core
+
+
+def _ensure_init(sanity_check: bool = False):
+    """Return the core, running PBookCore.init() exactly once per process.
+
+    The REPL calls this once upfront with sanity_check=True; every command
+    function then calls it again with the default before touching the core,
+    which is a no-op once already initialized - so commands stay correct
+    whether they run once (batch mode) or repeatedly (REPL session).
+    """
+    global _core_inited
+    core = _get_core()
+    if not _core_inited:
+        core.init(sanity_check=sanity_check)
+        _core_inited = True
+    return core
 
 
 def _catch_sig(sig, frame):
@@ -129,67 +219,96 @@ def _catch_sig(sig, frame):
     commands_get_output(f"kill -9 -- -{os.getpgrp()}")
 
 
-# ─── Interactive REPL namespace ───────────────────────────────────────────────
+# ─── REPL namespace & completion ──────────────────────────────────────────────
 
-# ─── REPL kwarg completer ─────────────────────────────────────────────────────
 
-_FUNC_KWARGS: dict[str, list[str]] = {
-    "show": ["username", "limit", "taskname", "days", "jeditaskid", "reqid", "status", "superstatus", "format"],
-    "showl": ["username", "limit", "taskname", "days", "jeditaskid", "reqid", "status", "superstatus"],
-    "kill": [],
-    "finish": ["soft"],
-    "retry": [
-        "newOpts",
-        "days",
-        "limit",
-        "site",
-        "excludedSite",
-        "includedSite",
-        "nFilesPerJob",
-        "nMaxFilesPerJob",
-        "nGBPerJob",
-        "nFiles",
-        "nEvents",
-        "loopingCheck",
-        "memory",
-        "avoidVP",
-        "ignoreMissingInDS",
-        "forceStaged",
-        "maxCore",
-    ],
-    "debug": ["modeOn"],
-    "get_user_job_metadata": [],
-    "recover_lost_files": ["test_mode"],
-    "set_secret": ["is_file"],
-    "list_secrets": ["full"],
-}
+def _build_namespace() -> dict:
+    """The REPL namespace: every registered Typer command, keyed by its real Python name."""
+    return {info.callback.__name__: info.callback for info in app.registered_commands}
 
-_KWARG_VALUES: dict[str, dict[str, list[str]]] = {
-    "show": {
-        "format": ["standard", "long", "json", "plain"],
-    },
-    "finish": {
-        "soft": ["True", "False"],
-    },
-    "debug": {
-        "modeOn": ["True", "False"],
-    },
-    "recover_lost_files": {
-        "test_mode": ["True", "False"],
-    },
-    "list_secrets": {
-        "full": ["True", "False"],
-    },
-    "set_secret": {
-        "is_file": ["True", "False"],
-    },
-    "retry": {
-        "loopingCheck": ["True", "False"],
-        "avoidVP": ["True", "False"],
-        "ignoreMissingInDS": ["True", "False"],
-        "forceStaged": ["True", "False"],
-    },
-}
+
+def _kwarg_names(func) -> list:
+    """All parameter names a function accepts - candidates for `name=` completion."""
+    return list(signature(func).parameters)
+
+
+def _kwarg_choices(func, name: str) -> list:
+    """Value choices for a parameter, derived from its type hint: Literal[...] members or True/False for bool.
+
+    Returned bare (unquoted) - readline's own quote-matching auto-closes an opening quote
+    the user already typed, so we don't need to (and shouldn't try to) add one ourselves.
+    """
+    try:
+        hints = get_type_hints(func, include_extras=True)
+    except Exception:
+        return []
+    ann = hints.get(name)
+    if ann is None:
+        return []
+    while hasattr(ann, "__metadata__"):
+        ann = ann.__origin__
+    if get_origin(ann) is Union:
+        non_none = [a for a in get_args(ann) if a is not type(None)]
+        if len(non_none) == 1:
+            ann = non_none[0]
+    if get_origin(ann) is Literal:
+        return [str(v) for v in get_args(ann)]
+    if ann is bool:
+        return ["True", "False"]
+    return []
+
+
+class _PBookCompleter:
+    """Readline completer: kwarg names and values when inside a call, names otherwise."""
+
+    def __init__(self, ns: dict) -> None:
+        self._ns = ns
+        self._base = rlcompleter.Completer(ns)
+        self._matches: list = []
+
+    def complete(self, text: str, state: int) -> Optional[str]:
+        if state == 0:
+            self._matches = self._compute(text)
+        return self._matches[state] if state < len(self._matches) else None
+
+    def _compute(self, text: str) -> list:
+        line = readline.get_line_buffer()
+
+        # Kwarg value completion (tier 1): last token is  kwarg=  or  kwarg='partial
+        # Return bare values - readline's own quote-matching auto-closes an opening quote
+        # the user already typed, so we deliberately don't add quotes ourselves here.
+        m_val = re.search(r"\b(\w+)\s*=\s*(['\"]?)(\w*)$", line)
+        m_func = re.match(r"(\w+)\s*\(", line)
+        if m_val and m_func:
+            kwarg, partial = m_val.group(1), m_val.group(3)
+            func = self._ns.get(m_func.group(1))
+            if func is not None:
+                # We're unambiguously past a `kwarg=` - this is a value position, not a
+                # name position, even if this particular kwarg has no enumerable choices
+                # (e.g. limit: int). Return here regardless, so an empty result doesn't
+                # fall through to tier 2's kwarg-name completion.
+                return [v for v in _kwarg_choices(func, kwarg) if v.startswith(partial)]
+
+        # Kwarg name completion (tier 2): cursor is inside an open call
+        m = re.search(r"(\w+)\s*\([^)]*$", line)
+        if m:
+            func = self._ns.get(m.group(1))
+            if func is not None:
+                hits = [k for k in _kwarg_names(func) if k.startswith(text)]
+                if hits:
+                    return hits
+
+        # Plain name completion (tier 3, rlcompleter fallback)
+        if not text:
+            # rlcompleter.complete() special-cases blank text by calling readline.insert_text()
+            # itself, which re-enters readline from inside this callback and confuses the active
+            # Tab press; list the namespace directly instead of delegating to it here
+            return sorted(k for k in self._base.namespace if not k.startswith("_"))
+        results, i = [], 0
+        while (c := self._base.complete(text, i)) is not None:
+            results.append(c.rstrip("()").rstrip("("))
+            i += 1
+        return results
 
 
 def _run_repl(ns: dict, banner: str) -> None:
@@ -219,308 +338,6 @@ def _run_repl(ns: dict, banner: str) -> None:
         more = console.push(line)
 
 
-class _PBookCompleter:
-    """Readline completer: kwarg names and values when inside a call, names otherwise."""
-
-    def __init__(self, ns: dict) -> None:
-        self._base = rlcompleter.Completer(ns)
-        self._matches: list[str] = []
-
-    def complete(self, text: str, state: int) -> Optional[str]:
-        if state == 0:
-            self._matches = self._compute(text)
-        return self._matches[state] if state < len(self._matches) else None
-
-    def _compute(self, text: str) -> list[str]:
-        line = readline.get_line_buffer()
-
-        # Kwarg value completion(tier 1, highest priority): last token is  kwarg=  or  kwarg='partial
-        # >>> show(format='j│
-        # Now the line matches both m_func (show() and m_val (line 234's regex: kwarg="format", quote="'", partial="j").
-        # It looks up _KWARG_VALUES["show"]["format"] (line 165) = ["standard", "long", "json", "plain"], filters to those starting with j → json.
-        # This branch is checked before tier 2, so once you're past the =, you get value suggestions instead of more kwarg names.
-        m_val = re.search(r"\b(\w+)\s*=\s*(['\"]?)(\w*)$", line)
-        m_func = re.match(r"(\w+)\s*\(", line)
-        if m_val and m_func:
-            kwarg, quote, partial = m_val.group(1), m_val.group(2), m_val.group(3)
-            func_name = m_func.group(1)
-            values = _KWARG_VALUES.get(func_name, {}).get(kwarg, [])
-            hits = [v for v in values if v.startswith(partial)]
-            if hits:
-                return hits
-
-        # Kwarg name completion: cursor is inside an open call
-        # >>> show(│
-        # readline.get_line_buffer() returns "show(". The regex on line 245, (\w+)\s*\([^)]*$, matches with func_name = "show" and nothing after the (.
-        # So it looks up _FUNC_KWARGS["show"] (line 133) and lists username, limit, taskname, days, jeditaskid, reqid, status, superstatus, format —
-        # rlcompleter is never consulted here.
-
-        # >>> show(form│
-        # Same regex still matches (func_name = "show", and [^)]* swallows form), but we filter _FUNC_KWARGS["show"] down to names starting with
-        # form → just format.
-        m = re.search(r"(\w+)\s*\([^)]*$", line)
-        if m:
-            hits = [k for k in _FUNC_KWARGS.get(m.group(1), []) if k.startswith(text)]
-            if hits:
-                return hits
-
-        #  Plain name completion (tier 3, the rlcompleter fallback): standard name completion, stripping trailing '(' rlcompleter adds to callables
-        # >>> sho│
-        # Not inside any (...), so tiers 1 and 2 don't match. Falls through to rlcompleter (line 257-260), which scans ns for names starting with
-        # sho → matches show, showl. Since these are callables, rlcompleter.complete() would normally return "show("/"showl("; the .rstrip("()").rstrip("(")
-        # on line 259 strips that back to plain show, showl.
-        if not text:
-            # rlcompleter.complete() special-cases blank text by calling readline.insert_text()
-            # itself, which re-enters readline from inside this callback and confuses the active
-            # Tab press; list the namespace directly instead of delegating to it here
-            return sorted(k for k in self._base.namespace if not k.startswith("_"))
-        results, i = [], 0
-        while (c := self._base.complete(text, i)) is not None:
-            results.append(c.rstrip("()").rstrip("("))
-            i += 1
-        return results
-
-
-_RETRY_ALLOWED_OPTS = [
-    "site",
-    "excludedSite",
-    "includedSite",
-    "nFilesPerJob",
-    "nMaxFilesPerJob",
-    "nGBPerJob",
-    "nFiles",
-    "nEvents",
-    "loopingCheck",
-    "maxNFilesPerJob",
-    "memory",
-    "ramCount",
-    "avoidVP",
-    "ignoreMissingInDS",
-    "forceStaged",
-    "maxCore",
-]
-
-
-def _build_namespace(core) -> dict:
-    _console = Console()
-
-    def help(command=None):
-        """Show available commands, or detailed help for a specific command."""
-        if command is not None:
-            name = command if isinstance(command, str) else command.__name__
-            func = ns.get(name, command if callable(command) else None)
-            if func is None:
-                _console.print(f"[red]Unknown command:[/red] {name}")
-                return
-            sig = str(signature(func)).replace("(", f"[bold cyan]{name}[/bold cyan](", 1)
-            _console.print(f"\n[bold]{sig}[/bold]")
-            doc = (func.__doc__ or "No description.").strip()
-            _console.print(f"\n{doc}\n")
-            return
-
-        # No argument - summary table
-        table = Table(box=box.SIMPLE, show_header=True, header_style="bold magenta")
-        table.add_column("Command", style="bold cyan", no_wrap=True)
-        table.add_column("Signature", style="dim", no_wrap=True)
-        table.add_column("Description")
-
-        _GROUPS = [
-            ("Tasks", ["show", "showl", "kill", "finish", "retry", "debug"]),
-            ("Files & input", ["get_user_job_metadata", "recover_lost_files", "reload_input"]),
-            ("Workflows", ["show_workflow", "kill_workflow", "retry_workflow", "finish_workflow", "pause_workflow", "resume_workflow"]),
-            ("Secrets", ["set_secret", "list_secrets", "delete_secret", "delete_all_secrets"]),
-            ("Auth", ["generate_credential"]),
-        ]
-        for group, names in _GROUPS:
-            table.add_section()
-            table.add_row(f"[bold white]{group}[/bold white]", "", "")
-            for name in names:
-                func = ns.get(name)
-                if func is None:
-                    continue
-                sig = str(signature(func))
-                doc = (func.__doc__ or "").strip().splitlines()[0]
-                table.add_row(f"  {name}", sig, doc)
-
-        _console.print(table)
-        _console.print("Usage: [bold]help(show)[/bold]  or  [bold]pbook show --help[/bold]\n")
-
-    def show(taskID=None, *, username=None, limit=1000, taskname=None, days=14, jeditaskid=None, reqid=None, status=None, superstatus=None, format="standard"):
-        """Print task records.
-
-        taskID: jediTaskID / reqID / 'run' (active) / 'fin' (terminated) / omit for all.
-        format: standard | long | json | plain
-        """
-        kwargs = {
-            k: v
-            for k, v in dict(
-                username=username,
-                limit=limit,
-                taskname=taskname,
-                days=days,
-                jeditaskid=jeditaskid,
-                reqid=reqid,
-                status=status,
-                superstatus=superstatus,
-                format=format,
-            ).items()
-            if v is not None
-        }
-        kwargs.setdefault("limit", limit)
-        kwargs.setdefault("days", days)
-        kwargs["format"] = format
-        return core.show(taskID, **kwargs) if taskID is not None else core.show(**kwargs)
-
-    def showl(taskID=None, *, username=None, limit=1000, taskname=None, days=14, jeditaskid=None, reqid=None, status=None, superstatus=None):
-        """Print task records in long format (shortcut for show(..., format='long'))."""
-        return show(
-            taskID,
-            username=username,
-            limit=limit,
-            taskname=taskname,
-            days=days,
-            jeditaskid=jeditaskid,
-            reqid=reqid,
-            status=status,
-            superstatus=superstatus,
-            format="long",
-        )
-
-    def kill(taskIDs):
-        """Kill tasks. taskIDs: int, list of ints, or 'all' for all active tasks."""
-        if taskIDs == "all":
-            return _parallel(lambda t: core.kill(t.jeditaskid), core.get_active_tasks())
-        elif isinstance(taskIDs, (list, tuple)):
-            return _parallel(core.kill, taskIDs)
-        elif isinstance(taskIDs, int):
-            return [core.kill(taskIDs)]
-
-        print("Error: Invalid argument")
-        return None
-
-    def finish(taskIDs, soft=False):
-        """Finish tasks. taskIDs: int, [int,...], or 'all' for all active tasks. soft=True waits for running jobs."""
-        if taskIDs == "all":
-            return _parallel(
-                lambda t: core.finish.original_func(core, t.jeditaskid, soft=soft),
-                core.get_active_tasks(),
-            )
-        elif isinstance(taskIDs, (list, tuple)):
-            return _parallel(lambda tid: core.finish(tid, soft=soft), taskIDs)
-        elif isinstance(taskIDs, int):
-            return [core.finish(taskIDs, soft=soft)]
-
-        print("Error: Invalid argument")
-        return None
-
-    def retry(taskIDs, newOpts=None, days=14, limit=1000, **kwargs):
-        """Retry failed/cancelled tasks.
-
-        taskIDs (required): int, list of ints, or 'all'.
-        Allowed kwargs: site, excludedSite, includedSite, nFilesPerJob, nMaxFilesPerJob,
-          nGBPerJob, nFiles, nEvents, loopingCheck, memory, avoidVP,
-          ignoreMissingInDS, forceStaged, maxCore.
-        Example: retry('all', loopingCheck=True)
-        """
-        if newOpts is None:
-            newOpts = dict(kwargs)
-        for key in list(newOpts):
-            if key == "memory":
-                newOpts["ramCount"] = newOpts.pop(key)
-            elif key == "maxCore":
-                newOpts["maxCoreCount"] = newOpts.pop(key)
-            elif key not in _RETRY_ALLOWED_OPTS:
-                print(f'Error: Unknown option "{key}"')
-                return None
-        opts = newOpts or None
-        if isinstance(taskIDs, (list, tuple)):
-            return _parallel(lambda tid: core.retry(tid, newOpts=opts), taskIDs)
-        elif isinstance(taskIDs, int):
-            return [core.retry(taskIDs, newOpts=opts)]
-        elif taskIDs == "all":
-            data = core.show(status="finished", days=days, limit=limit, format="json")
-            return _parallel(lambda d: core.retry.original_func(core, d["jediTaskID"], newOpts=opts), data)
-
-        print("Error: Invalid argument")
-        return None
-
-    def debug(PandaID, modeOn):
-        """Toggle debug mode for a subjob. modeOn: True/False."""
-        core.debug(PandaID, modeOn)
-
-    def get_user_job_metadata(taskID, outputFileName):
-        """Write user metadata of successful jobs to a JSON file."""
-        core.getUserJobMetadata(taskID, outputFileName)
-
-    def reload_input(task_id):
-        """Reload input dataset and retry with new contents."""
-        core.reload_input(task_id)
-
-    def recover_lost_files(taskID, test_mode=False):
-        """Request recovery of lost files from a task."""
-        core.recover_lost_files(taskID, test_mode)
-
-    def show_workflow(request_id):
-        """Show workflow status."""
-        _, output = core.execute_workflow_command("get_status", request_id)
-        if output:
-            print(output)
-
-    def kill_workflow(request_id):
-        """Kill a workflow."""
-        _, output = core.execute_workflow_command("abort", request_id)
-        if output:
-            print(output[0][-1])
-
-    def retry_workflow(request_id):
-        """Retry a workflow."""
-        _, output = core.execute_workflow_command("retry", request_id)
-        if output:
-            print(output[0][-1])
-
-    def finish_workflow(request_id):
-        """Finish a workflow."""
-        _, output = core.execute_workflow_command("finish", request_id)
-        if output:
-            print(output[0][-1])
-
-    def pause_workflow(request_id):
-        """Pause a workflow."""
-        _, output = core.execute_workflow_command("suspend", request_id)
-        if output:
-            print(output[0][-1])
-
-    def resume_workflow(request_id):
-        """Resume a workflow."""
-        _, output = core.execute_workflow_command("resume", request_id)
-        if output:
-            print(output[0][-1])
-
-    def set_secret(key, value, is_file=False):
-        """Set a secret key-value pair. is_file=True to upload a file."""
-        core.set_secret(key, value, is_file)
-
-    def delete_secret(key):
-        """Delete a secret."""
-        core.set_secret(key, None)
-
-    def delete_all_secrets():
-        """Delete all secrets."""
-        core.set_secret(None, None)
-
-    def list_secrets(full=False):
-        """List secrets. full=True to show full values."""
-        core.list_secrets(full)
-
-    def generate_credential():
-        """Generate a new proxy or token."""
-        core.generate_credential()
-
-    # Generate the namespace with the local functions and exclude any imported functions or variables
-    ns = {k: v for k, v in locals().items() if callable(v) and getattr(v, "__module__", None) == __name__}
-    return ns
-
-
 # ─── Top-level callback ───────────────────────────────────────────────────────
 
 
@@ -533,13 +350,13 @@ def _main(
     dev_srv: bool = typer.Option(False, "--devSrv", hidden=True),
     intr_srv: bool = typer.Option(False, "--intrSrv", hidden=True),
     prompt_with_newline: bool = typer.Option(False, "--prompt_with_newline", hidden=True),
+    python3: bool = typer.Option(False, "-3", hidden=True),
 ) -> None:
     """PanDA task bookkeeper. Run without arguments for interactive mode."""
     if version:
         typer.echo(f"Version: {PandaToolsPkgInfo.release_version}")
         raise typer.Exit()
 
-    # Set up the development or integration server
     if dev_srv:
         Client.useDevServer()
     if intr_srv:
@@ -555,31 +372,25 @@ def _main(
     global _fork_child_pid
     _fork_child_pid = os.fork()
 
-    # Fork failed
     if _fork_child_pid == -1:
         typer.echo("ERROR: Failed to fork", err=True)
         raise typer.Exit(1)
 
-    # Child process
     if _fork_child_pid == 0:
         if verbose:
             typer.echo(str(ctx.params))
         if prompt_with_newline:
             sys.ps1 = ">>> \n"
-        core = _make_core(verbose)
-        ns = _build_namespace(core)
+        _ensure_init(sanity_check=True)
+        ns = _build_namespace()
 
-        # The user wants to execute a Python code snippet instead of entering the REPL
         if command_string:
-            core.init()
             exec(command_string, {}, ns)  # noqa: S102
             from pandaclient import PBookCore as _PBC
 
             raise typer.Exit(0 if _PBC.func_return_value else 1)
-        core.init()
         _run_repl(ns, banner=f"\nStart pBook {PandaToolsPkgInfo.release_version}")
 
-    # Parent process
     else:
         signal.signal(signal.SIGINT, _catch_sig)
         signal.signal(signal.SIGHUP, _catch_sig)
@@ -592,219 +403,388 @@ def _main(
         raise typer.Exit(0)
 
 
-# ─── Subcommands ──────────────────────────────────────────────────────────────
+# ─── Commands ──────────────────────────────────────────────────────────────────
+
+_HELP_GROUPS = [
+    ("Tasks", ["show", "showl", "kill", "finish", "retry", "debug"]),
+    ("Files & input", ["get_user_job_metadata", "recover_lost_files", "reload_input"]),
+    ("Workflows", ["show_workflow", "kill_workflow", "retry_workflow", "finish_workflow", "pause_workflow", "resume_workflow"]),
+    ("Secrets", ["set_secret", "list_secrets", "delete_secret", "delete_all_secrets"]),
+    ("Auth", ["generate_credential"]),
+]
+
+
+def _type_name(ann) -> str:
+    """Render a resolved type annotation as a short, human-readable name (Optional[str], Literal[...], etc.)."""
+    if ann is None or ann is type(None):
+        return ""
+    origin = get_origin(ann)
+    if origin is Union:
+        args = get_args(ann)
+        non_none = [a for a in args if a is not type(None)]
+        if len(non_none) == 1 and len(args) == 2:
+            return f"Optional[{_type_name(non_none[0])}]"
+        return " | ".join(_type_name(a) for a in args)
+    if origin is Literal:
+        return "Literal[" + ", ".join(repr(v) for v in get_args(ann)) + "]"
+    if origin is not None:
+        args = get_args(ann)
+        origin_name = getattr(origin, "__name__", str(origin))
+        return f"{origin_name}[{', '.join(_type_name(a) for a in args)}]" if args else origin_name
+    return getattr(ann, "__name__", str(ann))
+
+
+def _format_signature(func) -> str:
+    """A clean '(param: Type = default, ...)' string, stripping the Typer/Annotated plumbing."""
+    try:
+        hints = get_type_hints(func, include_extras=True)
+    except Exception:
+        hints = {}
+    parts = []
+    for pname, p in signature(func).parameters.items():
+        ann = hints.get(pname)
+        while hasattr(ann, "__metadata__"):
+            ann = ann.__origin__
+        piece = pname
+        type_str = _type_name(ann)
+        if type_str:
+            piece += f": {type_str}"
+        if p.default is not Parameter.empty:
+            piece += f" = {p.default!r}"
+        parts.append(piece)
+    return f"({', '.join(parts)})"
+
+
+@app.command()
+def help(
+    command: Annotated[Optional[str], typer.Argument(help="Command name for detailed help")] = None,
+) -> None:
+    """Show available commands, or detailed help for a specific command."""
+    ns = _build_namespace()
+    console = Console()
+
+    if command is not None:
+        name = command if isinstance(command, str) else command.__name__
+        func = ns.get(name, command if callable(command) else None)
+        if func is None:
+            console.print(f"[red]Unknown command:[/red] {_esc(name)}")
+            return
+        console.print(f"\n[bold cyan]{name}[/bold cyan][bold]{_esc(_format_signature(func))}[/bold]")
+        doc = (func.__doc__ or "No description.").strip()
+        console.print(f"\n{_esc(doc)}\n")
+        return
+
+    table = Table(box=box.SIMPLE, show_header=True, header_style="bold magenta")
+    table.add_column("Command", style="bold cyan", no_wrap=True)
+    table.add_column("Description")
+    table.add_column("Signature")
+
+    for group, names in _HELP_GROUPS:
+        table.add_section()
+        table.add_row(f"[bold white]{group}[/bold white]", "", "")
+        for name in names:
+            func = ns.get(name)
+            if func is None:
+                continue
+            doc = (func.__doc__ or "").strip().splitlines()[0] if func.__doc__ else ""
+            sig = console.highlighter(_format_signature(func))
+            table.add_row(f"  {name}", _esc(doc), sig)
+
+    console.print(table)
+    console.print("Usage: [bold]help(show)[/bold]  or  [bold]pbook show --help[/bold]\n")
 
 
 @app.command()
 def show(
-    task_id: Optional[str] = typer.Argument(None, help="jediTaskID, reqID, 'run' (active only), or 'fin' (terminated only)"),
-    username: Optional[str] = typer.Option(None, "--username", help="Filter by username"),
-    limit: int = typer.Option(1000, "--limit", help="Maximum number of records"),
-    taskname: Optional[str] = typer.Option(None, "--taskname", help="Filter by task name"),
-    days: int = typer.Option(14, "--days", help="Look back N days (capped at 90 without a task ID)"),
-    jeditaskid: Optional[int] = typer.Option(None, "--jeditaskid", help="Filter by jediTaskID"),
-    reqid: Optional[int] = typer.Option(None, "--reqid", help="Filter by reqID"),
-    status: Optional[str] = typer.Option(None, "--status", help="Filter by task status"),
-    superstatus: Optional[str] = typer.Option(None, "--superstatus", help="Filter by super-status"),
-    output_format: str = typer.Option("standard", "--format", help="Output format: standard|long|json|plain"),
+    task_id: Annotated[Optional[str], typer.Argument(help="jediTaskID, reqID, 'run' (active only), or 'fin' (terminated only)")] = None,
+    username: Annotated[Optional[str], typer.Option(help="Filter by username")] = None,
+    limit: Annotated[int, typer.Option(help="Maximum number of records")] = 1000,
+    taskname: Annotated[Optional[str], typer.Option(help="Filter by task name")] = None,
+    days: Annotated[int, typer.Option(help="Look back N days (capped at 90 without a task ID)")] = 14,
+    jeditaskid: Annotated[Optional[int], typer.Option(help="Filter by jediTaskID")] = None,
+    reqid: Annotated[Optional[int], typer.Option(help="Filter by reqID")] = None,
+    status: Annotated[Optional[str], typer.Option(help="Filter by task status")] = None,
+    superstatus: Annotated[Optional[str], typer.Option(help="Filter by super-status")] = None,
+    format: Annotated[Literal["standard", "long", "json", "plain"], typer.Option("--format", help="Output format")] = "standard",
 ) -> None:
-    """Print task records."""
-    core = _get_core()
-    core.init(sanity_check=False)
-    kwargs: dict = {"limit": limit, "days": days, "format": output_format}
-    for k, v in [
-        ("username", username),
-        ("taskname", taskname),
-        ("jeditaskid", jeditaskid),
-        ("reqid", reqid),
-        ("status", status),
-        ("superstatus", superstatus),
-    ]:
-        if v is not None:
-            kwargs[k] = v
+    """Print task records.
+
+    The first argument (task_id) can be a jediTaskID or reqID, or 'run' (show active tasks
+    only), or 'fin' (show terminated tasks only), or can be omitted. Records are fetched
+    directly from the PanDA server, so they are always up to date. Note that days is capped
+    at 90 days unless a jediTaskID or reqID is specified, in which case tasks of any age are
+    returned. The default filter conditions are: username=(name from user voms proxy),
+    limit=1000, days=14, format='standard'.
+
+    example:
+      >>> show()
+      >>> show(123)
+      >>> show(12345678, format='long')
+      >>> show(taskname='my_task_name')
+      >>> show('run')
+      >>> show('fin', days=7, limit=100)
+      >>> show(format='json')
+    """
+    core = _ensure_init()
+    kwargs = {
+        k: v
+        for k, v in dict(
+            username=username,
+            limit=limit,
+            taskname=taskname,
+            days=days,
+            jeditaskid=jeditaskid,
+            reqid=reqid,
+            status=status,
+            superstatus=superstatus,
+        ).items()
+        if v is not None
+    }
+    kwargs["format"] = format
     if task_id is not None:
         try:
             first_arg = int(task_id)
-        except ValueError:
+        except (TypeError, ValueError):
             first_arg = task_id
-        core.show(first_arg, **kwargs)
-    else:
-        core.show(**kwargs)
+        return core.show(first_arg, **kwargs)
+    return core.show(**kwargs)
 
 
 @app.command()
 def showl(
-    task_id: Optional[str] = typer.Argument(None, help="jediTaskID, reqID, 'run', or 'fin'"),
-    username: Optional[str] = typer.Option(None, "--username"),
-    limit: int = typer.Option(1000, "--limit"),
-    taskname: Optional[str] = typer.Option(None, "--taskname"),
-    days: int = typer.Option(14, "--days"),
-    jeditaskid: Optional[int] = typer.Option(None, "--jeditaskid"),
-    reqid: Optional[int] = typer.Option(None, "--reqid"),
-    status: Optional[str] = typer.Option(None, "--status"),
-    superstatus: Optional[str] = typer.Option(None, "--superstatus"),
+    task_id: Annotated[Optional[str], typer.Argument(help="jediTaskID, reqID, 'run', or 'fin'")] = None,
+    username: Annotated[Optional[str], typer.Option(help="Filter by username")] = None,
+    limit: Annotated[int, typer.Option(help="Maximum number of records")] = 1000,
+    taskname: Annotated[Optional[str], typer.Option(help="Filter by task name")] = None,
+    days: Annotated[int, typer.Option(help="Look back N days (capped at 90 without a task ID)")] = 14,
+    jeditaskid: Annotated[Optional[int], typer.Option(help="Filter by jediTaskID")] = None,
+    reqid: Annotated[Optional[int], typer.Option(help="Filter by reqID")] = None,
+    status: Annotated[Optional[str], typer.Option(help="Filter by task status")] = None,
+    superstatus: Annotated[Optional[str], typer.Option(help="Filter by super-status")] = None,
 ) -> None:
-    """Print task records in long format (shortcut for show --format long)."""
-    core = _get_core()
-    core.init(sanity_check=False)
-    kwargs: dict = {"limit": limit, "days": days, "format": "long"}
-    for k, v in [
-        ("username", username),
-        ("taskname", taskname),
-        ("jeditaskid", jeditaskid),
-        ("reqid", reqid),
-        ("status", status),
-        ("superstatus", superstatus),
-    ]:
-        if v is not None:
-            kwargs[k] = v
-    if task_id is not None:
-        try:
-            first_arg = int(task_id)
-        except ValueError:
-            first_arg = task_id
-        core.show(first_arg, **kwargs)
-    else:
-        core.show(**kwargs)
+    """Print task records in long format (shortcut for show --format long).
+
+    See help(show) for the available filter keywords.
+
+    example:
+      >>> showl()
+      >>> showl(123)
+      >>> showl(12345678)
+      >>> showl(taskname='my_task_name')
+    """
+    return show(
+        task_id,
+        username=username,
+        limit=limit,
+        taskname=taskname,
+        days=days,
+        jeditaskid=jeditaskid,
+        reqid=reqid,
+        status=status,
+        superstatus=superstatus,
+        format="long",
+    )
 
 
 @app.command()
 def kill(
-    task_ids: str = typer.Argument(..., help="Task ID, comma-separated IDs, or 'all'"),
+    task_ids: Annotated[str, typer.Argument(help="Task ID, comma-separated IDs, or 'all'")],
 ) -> None:
-    """Kill tasks."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Kill tasks.
+
+    Kill all subJobs in task_ids (ID or a list of IDs, can be either jediTaskID or reqID).
+    If 'all', kill all active tasks of the user.
+
+    example:
+      >>> kill(123)
+      >>> kill([123, 345, 567])
+      >>> kill('all')
+    """
+    core = _ensure_init()
     ids = _parse_ids(task_ids)
     if ids == "all":
-        _parallel(lambda t: core.kill(t.jeditaskid), core.get_active_tasks())
+        return _parallel(lambda t: core.kill(t.jeditaskid), core.get_active_tasks())
     elif isinstance(ids, list):
-        _parallel(core.kill, ids)
-    else:
-        core.kill(ids)
+        return _parallel(core.kill, ids)
+    return core.kill(ids)
 
 
 @app.command()
 def finish(
-    task_ids: str = typer.Argument(..., help="Task ID, comma-separated IDs, or 'all'"),
-    soft: bool = typer.Option(False, "--soft", help="Wait for running jobs to finish instead of killing them"),
+    task_ids: Annotated[str, typer.Argument(help="Task ID, comma-separated IDs, or 'all'")],
+    soft: Annotated[bool, typer.Option("--soft", help="Wait for running jobs to finish instead of killing them")] = False,
 ) -> None:
-    """Finish tasks."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Finish tasks.
+
+    Finish all subJobs in task_ids (ID or a list of IDs, can be either jediTaskID or reqID).
+    If task_ids is 'all', finish all active tasks of the user. If soft is False (default),
+    all running jobs are killed and the task finishes immediately. If soft is True, new jobs
+    are not generated and the task finishes once all running jobs finish.
+
+    example:
+      >>> finish(123)
+      >>> finish(234, soft=True)
+      >>> finish([123, 345, 567])
+      >>> finish('all')
+    """
+    core = _ensure_init()
     ids = _parse_ids(task_ids)
     if ids == "all":
-        _parallel(lambda t: core.finish.original_func(core, t.jeditaskid, soft=soft), core.get_active_tasks())
+        return _parallel(lambda t: core.finish.original_func(core, t.jeditaskid, soft=soft), core.get_active_tasks())
     elif isinstance(ids, list):
-        _parallel(lambda tid: core.finish(tid, soft=soft), ids)
-    else:
-        core.finish(ids, soft=soft)
+        return _parallel(lambda tid: core.finish(tid, soft=soft), ids)
+    return core.finish(ids, soft=soft)
 
 
 @app.command()
 def retry(
-    task_ids: str = typer.Argument(..., help="Task ID, comma-separated IDs, or 'all'"),
-    days: int = typer.Option(14, "--days", help="Look-back window when task_ids='all'"),
-    limit: int = typer.Option(1000, "--limit", help="Max tasks to retry when task_ids='all'"),
-    site: Optional[str] = typer.Option(None, "--site"),
-    excluded_site: Optional[str] = typer.Option(None, "--excludedSite"),
-    included_site: Optional[str] = typer.Option(None, "--includedSite"),
-    n_files_per_job: Optional[int] = typer.Option(None, "--nFilesPerJob"),
-    n_max_files_per_job: Optional[int] = typer.Option(None, "--nMaxFilesPerJob"),
-    n_gb_per_job: Optional[float] = typer.Option(None, "--nGBPerJob"),
-    n_files: Optional[int] = typer.Option(None, "--nFiles"),
-    n_events: Optional[int] = typer.Option(None, "--nEvents"),
-    looping_check: Optional[bool] = typer.Option(None, "--loopingCheck"),
-    memory: Optional[int] = typer.Option(None, "--memory"),
-    avoid_vp: Optional[bool] = typer.Option(None, "--avoidVP"),
-    ignore_missing_in_ds: Optional[bool] = typer.Option(None, "--ignoreMissingInDS"),
-    force_staged: Optional[bool] = typer.Option(None, "--forceStaged"),
-    max_core: Optional[int] = typer.Option(None, "--maxCore"),
+    task_ids: Annotated[str, typer.Argument(help="Task ID, comma-separated IDs, or 'all'")],
+    days: Annotated[int, typer.Option("--days", help="Look-back window when task_ids='all'")] = 14,
+    limit: Annotated[int, typer.Option("--limit", help="Max tasks to retry when task_ids='all'")] = 1000,
+    site: Annotated[Optional[str], typer.Option("--site")] = None,
+    excludedSite: Annotated[Optional[str], typer.Option("--excludedSite")] = None,
+    includedSite: Annotated[Optional[str], typer.Option("--includedSite")] = None,
+    nFilesPerJob: Annotated[Optional[int], typer.Option("--nFilesPerJob")] = None,
+    nMaxFilesPerJob: Annotated[Optional[int], typer.Option("--nMaxFilesPerJob")] = None,
+    nGBPerJob: Annotated[Optional[float], typer.Option("--nGBPerJob")] = None,
+    nFiles: Annotated[Optional[int], typer.Option("--nFiles")] = None,
+    nEvents: Annotated[Optional[int], typer.Option("--nEvents")] = None,
+    loopingCheck: Annotated[Optional[bool], typer.Option("--loopingCheck")] = None,
+    memory: Annotated[Optional[int], typer.Option("--memory")] = None,
+    avoidVP: Annotated[Optional[bool], typer.Option("--avoidVP")] = None,
+    ignoreMissingInDS: Annotated[Optional[bool], typer.Option("--ignoreMissingInDS")] = None,
+    forceStaged: Annotated[Optional[bool], typer.Option("--forceStaged")] = None,
+    maxCore: Annotated[Optional[int], typer.Option("--maxCore")] = None,
 ) -> None:
-    """Retry failed/cancelled tasks."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Retry failed/cancelled tasks.
+
+    Retry failed/cancelled subJobs in task_ids (ID or a list of IDs, can be either jediTaskID
+    or reqID). Allowed options to overwrite task parameters for new attempts: site,
+    excludedSite, includedSite, nFilesPerJob, nMaxFilesPerJob, nGBPerJob, nFiles, nEvents,
+    loopingCheck, memory, avoidVP, ignoreMissingInDS, forceStaged, maxCore. If input files
+    were used or are being used by other jobs for the same output dataset container, those
+    files are skipped to avoid job duplication when retrying failed subjobs.
+
+    If task_ids is 'all', it retries 1000 tasks at most that have finished for the last 14
+    days. It is possible to retry more tasks by setting the days and limit options. If
+    named arguments are specified, they are applied to all retried tasks.
+
+    example:
+      >>> retry(123)
+      >>> retry([123, 345, 567])
+      >>> retry(789, excludedSite='siteA,siteB')
+      >>> retry('all')
+      >>> retry('all', days=30, limit=2000)
+    """
+    core = _ensure_init()
     new_opts = {
         k: v
         for k, v in {
             "site": site,
-            "excludedSite": excluded_site,
-            "includedSite": included_site,
-            "nFilesPerJob": n_files_per_job,
-            "nMaxFilesPerJob": n_max_files_per_job,
-            "nGBPerJob": n_gb_per_job,
-            "nFiles": n_files,
-            "nEvents": n_events,
-            "loopingCheck": looping_check,
+            "excludedSite": excludedSite,
+            "includedSite": includedSite,
+            "nFilesPerJob": nFilesPerJob,
+            "nMaxFilesPerJob": nMaxFilesPerJob,
+            "nGBPerJob": nGBPerJob,
+            "nFiles": nFiles,
+            "nEvents": nEvents,
+            "loopingCheck": loopingCheck,
             "ramCount": memory,
-            "avoidVP": avoid_vp,
-            "ignoreMissingInDS": ignore_missing_in_ds,
-            "forceStaged": force_staged,
-            "maxCoreCount": max_core,
+            "avoidVP": avoidVP,
+            "ignoreMissingInDS": ignoreMissingInDS,
+            "forceStaged": forceStaged,
+            "maxCoreCount": maxCore,
         }.items()
         if v is not None
     }
     opts = new_opts or None
     ids = _parse_ids(task_ids)
     if isinstance(ids, list):
-        _parallel(lambda tid: core.retry(tid, newOpts=opts), ids)
-    elif isinstance(ids, int):
-        core.retry(ids, newOpts=opts)
-    else:
+        return _parallel(lambda tid: core.retry(tid, newOpts=opts), ids)
+    elif ids == "all":
         data = core.show(status="finished", days=days, limit=limit, format="json")
-        _parallel(lambda d: core.retry.original_func(core, d["jediTaskID"], newOpts=opts), data)
+        return _parallel(lambda d: core.retry.original_func(core, d["jediTaskID"], newOpts=opts), data)
+    return core.retry(ids, newOpts=opts)
 
 
 @app.command()
 def debug(
-    panda_id: int = typer.Argument(..., help="PanDA subjob ID"),
-    mode_on: bool = typer.Argument(..., help="True to enable, False to disable"),
+    panda_id: Annotated[int, typer.Argument(help="PanDA subjob ID")],
+    mode_on: Annotated[bool, typer.Argument(help="True to enable, False to disable")],
 ) -> None:
-    """Toggle debug mode for a subjob."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Toggle debug mode for a subjob.
+
+    mode_on is True/False to enable/disable the debug mode. Note that the maximum number of
+    debug subjobs is limited. If you already hit the limit you need to disable the debug mode
+    for a subjob before debugging another subjob.
+
+    example:
+      >>> debug(1234, True)
+    """
+    core = _ensure_init()
     core.debug(panda_id, mode_on)
 
 
 @app.command(name="get-user-job-metadata")
 def get_user_job_metadata(
-    task_id: int = typer.Argument(..., help="Task ID"),
-    output_file: str = typer.Argument(..., help="Output JSON file path"),
+    task_id: Annotated[int, typer.Argument(help="Task ID")],
+    output_file: Annotated[str, typer.Argument(help="Output JSON file path")],
 ) -> None:
-    """Write user metadata of successful jobs to a JSON file."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Write user metadata of successful jobs to a JSON file.
+
+    Get user metadata of successful jobs in a task and write them in a json file.
+
+    example:
+      >>> get_user_job_metadata(123, 'output.json')
+    """
+    core = _ensure_init()
     core.getUserJobMetadata(task_id, output_file)
 
 
 @app.command(name="reload-input")
 def reload_input(
-    task_id: int = typer.Argument(..., help="Task ID"),
+    task_id: Annotated[int, typer.Argument(help="Task ID")],
 ) -> None:
-    """Reload input dataset and retry the task with new contents."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Reload input dataset and retry the task with new contents.
+
+    This is useful when input dataset contents are changed after the task is submitted.
+
+    example:
+      >>> reload_input(123)
+    """
+    core = _ensure_init()
     core.reload_input(task_id)
 
 
 @app.command(name="recover-lost-files")
 def recover_lost_files(
-    task_id: int = typer.Argument(..., help="Task ID"),
-    test_mode: bool = typer.Option(False, "--test-mode", help="Dry-run mode"),
+    task_id: Annotated[int, typer.Argument(help="Task ID")],
+    test_mode: Annotated[bool, typer.Option("--test-mode", help="Dry-run mode")] = False,
 ) -> None:
-    """Request recovery of lost files from a task."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Request recovery of lost files from a task.
+
+    Send a request to recover lost files produced by a task. Set test_mode=True for testing.
+
+    example:
+      >>> recover_lost_files(123)
+      >>> recover_lost_files(123, test_mode=True)
+    """
+    core = _ensure_init()
     core.recover_lost_files(task_id, test_mode)
 
 
 @app.command(name="show-workflow")
 def show_workflow(
-    request_id: int = typer.Argument(..., help="Workflow request ID"),
+    request_id: Annotated[int, typer.Argument(help="Workflow request ID")],
 ) -> None:
-    """Show workflow status."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Show workflow status.
+
+    Send a request to show the status of a workflow.
+
+    example:
+      >>> show_workflow(456)
+    """
+    core = _ensure_init()
     _, output = core.execute_workflow_command("get_status", request_id)
     if output:
         print(output)
@@ -812,11 +792,16 @@ def show_workflow(
 
 @app.command(name="kill-workflow")
 def kill_workflow(
-    request_id: int = typer.Argument(..., help="Workflow request ID"),
+    request_id: Annotated[int, typer.Argument(help="Workflow request ID")],
 ) -> None:
-    """Kill a workflow."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Kill a workflow.
+
+    Send a request to kill a workflow.
+
+    example:
+      >>> kill_workflow(456)
+    """
+    core = _ensure_init()
     _, output = core.execute_workflow_command("abort", request_id)
     if output:
         print(output[0][-1])
@@ -824,11 +809,16 @@ def kill_workflow(
 
 @app.command(name="retry-workflow")
 def retry_workflow(
-    request_id: int = typer.Argument(..., help="Workflow request ID"),
+    request_id: Annotated[int, typer.Argument(help="Workflow request ID")],
 ) -> None:
-    """Retry a workflow."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Retry a workflow.
+
+    Send a request to retry a workflow.
+
+    example:
+      >>> retry_workflow(456)
+    """
+    core = _ensure_init()
     _, output = core.execute_workflow_command("retry", request_id)
     if output:
         print(output[0][-1])
@@ -836,11 +826,16 @@ def retry_workflow(
 
 @app.command(name="finish-workflow")
 def finish_workflow(
-    request_id: int = typer.Argument(..., help="Workflow request ID"),
+    request_id: Annotated[int, typer.Argument(help="Workflow request ID")],
 ) -> None:
-    """Finish a workflow."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Finish a workflow.
+
+    Send a request to finish a workflow.
+
+    example:
+      >>> finish_workflow(456)
+    """
+    core = _ensure_init()
     _, output = core.execute_workflow_command("finish", request_id)
     if output:
         print(output[0][-1])
@@ -848,11 +843,16 @@ def finish_workflow(
 
 @app.command(name="pause-workflow")
 def pause_workflow(
-    request_id: int = typer.Argument(..., help="Workflow request ID"),
+    request_id: Annotated[int, typer.Argument(help="Workflow request ID")],
 ) -> None:
-    """Pause a workflow."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Pause a workflow.
+
+    Send a request to pause a workflow.
+
+    example:
+      >>> pause_workflow(456)
+    """
+    core = _ensure_init()
     _, output = core.execute_workflow_command("suspend", request_id)
     if output:
         print(output[0][-1])
@@ -860,11 +860,16 @@ def pause_workflow(
 
 @app.command(name="resume-workflow")
 def resume_workflow(
-    request_id: int = typer.Argument(..., help="Workflow request ID"),
+    request_id: Annotated[int, typer.Argument(help="Workflow request ID")],
 ) -> None:
-    """Resume a workflow."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Resume a workflow.
+
+    Send a request to resume a workflow.
+
+    example:
+      >>> resume_workflow(456)
+    """
+    core = _ensure_init()
     _, output = core.execute_workflow_command("resume", request_id)
     if output:
         print(output[0][-1])
@@ -872,42 +877,57 @@ def resume_workflow(
 
 @app.command(name="set-secret")
 def set_secret(
-    key: str = typer.Argument(..., help="Secret key"),
-    value: str = typer.Argument(..., help="Secret value or file path"),
-    is_file: bool = typer.Option(False, "--is-file", help="Treat value as a file path to upload"),
+    key: Annotated[str, typer.Argument(help="Secret key")],
+    value: Annotated[str, typer.Argument(help="Secret value or file path")],
+    is_file: Annotated[bool, typer.Option("--is-file", help="Treat value as a file path to upload")] = False,
 ) -> None:
-    """Set a secret key-value pair."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Set a secret key-value pair.
+
+    Define a pair of secret key-value strings. The value can be a file path to upload a
+    secret file when is_file=True.
+
+    example:
+      >>> set_secret('mykey', 'myvalue')
+      >>> set_secret('mykey', '/path/to/file', is_file=True)
+    """
+    core = _ensure_init()
     core.set_secret(key, value, is_file)
-
-
-@app.command(name="list-secrets")
-def list_secrets(
-    full: bool = typer.Option(False, "--full", help="Show full secret values"),
-) -> None:
-    """List secrets."""
-    core = _get_core()
-    core.init(sanity_check=False)
-    core.list_secrets(full)
 
 
 @app.command(name="delete-secret")
 def delete_secret(
-    key: str = typer.Argument(..., help="Secret key to delete"),
+    key: Annotated[str, typer.Argument(help="Secret key to delete")],
 ) -> None:
-    """Delete a secret."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    """Delete a secret.
+
+    example:
+      >>> delete_secret('mykey')
+    """
+    core = _ensure_init()
     core.set_secret(key, None)
 
 
 @app.command(name="delete-all-secrets")
 def delete_all_secrets() -> None:
     """Delete all secrets."""
-    core = _get_core()
-    core.init(sanity_check=False)
+    core = _ensure_init()
     core.set_secret(None, None)
+
+
+@app.command(name="list-secrets")
+def list_secrets(
+    full: Annotated[bool, typer.Option("--full", help="Show full secret values")] = False,
+) -> None:
+    """List secrets.
+
+    Value strings are truncated by default. full=True to see entire strings.
+
+    example:
+      >>> list_secrets()
+      >>> list_secrets(full=True)
+    """
+    core = _ensure_init()
+    core.list_secrets(full)
 
 
 @app.command(name="generate-credential")
@@ -919,7 +939,65 @@ def generate_credential() -> None:
 
 # ─── Entry point ──────────────────────────────────────────────────────────────
 
+# Global options that consume the following argv token as their own value, so the
+# subcommand-token scan below can skip over both.
+_GLOBAL_VALUE_OPTS = {"-c"}
+
+
+def _rewrite_legacy_kwargs(argv: list) -> list:
+    """Rewrite legacy bare `key=value` batch args into `--key=value`.
+
+    The pre-Typer pbook batch mode accepted `pbook show format=long`; Click only
+    recognizes `--format=long`. Find the subcommand, look up its real option names via
+    introspection (never a separately maintained list), and rewrite any later bare
+    `key=value` token whose key matches one of them. Anything else - already-dashed
+    flags, positional args that happen to contain "=" - passes through untouched.
+    """
+    i = 0
+    while i < len(argv):
+        tok = argv[i]
+        if tok in _GLOBAL_VALUE_OPTS:
+            i += 2
+            continue
+        if tok.startswith("-"):
+            i += 1
+            continue
+        break
+    if i >= len(argv):
+        return argv
+
+    sub_cmd = typer.main.get_command(app).commands.get(argv[i])
+    if sub_cmd is None:
+        return argv
+
+    option_flags = {}
+    flag_only = set()
+    for param in sub_cmd.params:
+        flags = [o for o in getattr(param, "opts", []) if o.startswith("--")]
+        if flags:
+            option_flags[param.name] = flags[0]
+            if getattr(param, "is_flag", False):
+                flag_only.add(param.name)
+
+    rewritten = argv[: i + 1]
+    for tok in argv[i + 1 :]:
+        key, sep, value = tok.partition("=")
+        if sep and not tok.startswith("-") and key in option_flags:
+            flag = option_flags[key]
+            if key in flag_only:
+                # Click flag-style options (e.g. --soft) take no value at all; the legacy
+                # syntax passed an explicit True/False, so translate that into presence
+                # (truthy) or absence (falsy - same as the option's own default) instead.
+                if value.strip().lower() in ("true", "1", "yes"):
+                    rewritten.append(flag)
+                continue
+            rewritten.append(f"{flag}={value}")
+        else:
+            rewritten.append(tok)
+    return rewritten
+
 
 def main() -> None:
     sys.argv[0] = "pbook"
+    sys.argv[1:] = _rewrite_legacy_kwargs(sys.argv[1:])
     app()
